ci: configure cron tests
